+++ config-model/src/main/java/com/yahoo/vespa/model/container/PlatformBundles.java
@@ -146,8 +146,8 @@ public class PlatformBundles {
             com.yahoo.search.searchers.RateLimitingSearcher.class.getName(),
             com.yahoo.vespa.streamingvisitors.MetricsSearcher.class.getName(),
             com.yahoo.vespa.streamingvisitors.StreamingBackend.class.getName(),
-            ai.vespa.llm.search.LLMSearcher.class.getName(),
-            ai.vespa.llm.search.RAGSearcher.class.getName(),
+            ai.vespa.search.llm.LLMSearcher.class.getName(),
+            ai.vespa.search.llm.RAGSearcher.class.getName(),
             ai.vespa.llm.clients.OpenAI.class.getName()
     );
