+++ container-search/src/main/resources/configdefinitions/llm-searcher.def
@@ -1,5 +1,5 @@
 # Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package=ai.vespa.llm.search
+package=ai.vespa.search.llm
 
 # Query propertry prefix for options
 propertyPrefix string default="llm"
