+++ container-search/src/test/java/ai/vespa/search/llm/RAGSearcherTest.java
@@ -1,5 +1,5 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package ai.vespa.llm.search;
+package ai.vespa.search.llm;
 
 import ai.vespa.llm.LanguageModel;
 import com.yahoo.component.ComponentId;
